Move delete_translation view to the translations module.
