+++ frontend/public/static/locale/en-US/translate.ftl
@@ -443,6 +443,7 @@ notification--unable-to-approve-translation = Unable to approve translation
 notification--unable-to-unapprove-translation = Unable to unapprove translation
 notification--unable-to-reject-translation = Unable to reject translation
 notification--unable-to-unreject-translation = Unable to unreject translation
+notification--unable-to-delete-translation = Unable to delete translation
 notification--same-translation = Same translation already exists
 notification--tt-checks-enabled = Translate Toolkit Checks enabled
 notification--tt-checks-disabled = Translate Toolkit Checks disabled
