+++ pontoon/base/urls.py
@@ -64,11 +64,6 @@ urlpatterns = [
     url(r"^get-entities/", views.entities, name="pontoon.entities"),
     url(r"^perform-checks/", views.perform_checks, name="pontoon.perform.checks"),
     url(r"^get-history/", views.get_translation_history, name="pontoon.get_history"),
-    url(
-        r"^delete-translation/",
-        views.delete_translation,
-        name="pontoon.delete_translation",
-    ),
     url(r"^add-comment/", views.add_comment, name="pontoon.add_comment",),
     url(
         r"^other-locales/",
