+++ pontoon/translations/views.py
@@ -3,7 +3,9 @@ from __future__ import absolute_import
 from django.contrib.auth.decorators import login_required
 from django.db import transaction
 from django.http import JsonResponse
+from django.shortcuts import get_object_or_404
 from django.utils import timezone
+from django.utils.datastructures import MultiValueDictKeyError
 from django.views.decorators.http import require_POST
 
 from pontoon.actionlog.utils import log_action
@@ -22,6 +24,9 @@ from pontoon.translations import forms
 @login_required(redirect_field_name="", login_url="/403")
 @transaction.atomic
 def create_translation(request):
+    """
+    Create a new translation.
+    """
     form = forms.CreateTranslationForm(request.POST)
 
     if not form.is_valid():
@@ -114,3 +119,52 @@ def create_translation(request):
             "stats": TranslatedResource.objects.stats(project, resources, locale),
         }
     )
+
+
+@utils.require_AJAX
+@login_required(redirect_field_name="", login_url="/403")
+@transaction.atomic
+def delete_translation(request):
+    """Delete given translation."""
+    try:
+        translation_id = request.POST["translation"]
+    except MultiValueDictKeyError as e:
+        return JsonResponse(
+            {"status": False, "message": "Bad Request: {error}".format(error=e)},
+            status=400,
+        )
+
+    translation = get_object_or_404(Translation, pk=translation_id)
+    entity = translation.entity
+    project = entity.resource.project
+    locale = translation.locale
+
+    # Read-only translations cannot be deleted
+    if utils.readonly_exists(project, locale):
+        return JsonResponse(
+            {
+                "status": False,
+                "message": "Forbidden: This string is in read-only mode.",
+            },
+            status=403,
+        )
+
+    # Only privileged users or authors can delete translations
+    if not translation.rejected or not (
+        request.user.can_translate(locale, project)
+        or request.user == translation.user
+        or translation.approved
+    ):
+        return JsonResponse(
+            {
+                "status": False,
+                "message": "Forbidden: You can't delete this translation.",
+            },
+            status=403,
+        )
+
+    translation.delete()
+
+    log_action("translation:deleted", request.user, entity=entity, locale=locale)
+
+    return JsonResponse({"status": True})
